API

GnuplotLite.Gnuplot — Type

 struct Gnuplot

This struct is not exported, as it is not supposed to be created by a user, other than through calling gnuplot(), or save().

Messages can be sent to a Gnuplot instance using the send() method:

gnuplot() do g
    g |> send("plot sin(x)")
end

GnuplotLite.Msg — Type

 struct Msg

This wraps the actions that are created using the different variants of the send() method.

Messages can be composed (unlike ∘, going left to right) using the * operator.

Base.close — Method

 close(::Gnuplot)

Close the Gnuplot process by closing the underlying channel. The actual process my continue living if terminals are left open.

GnuplotLite.gnuplot — Method

 gnuplot(::Function; kwargs...)

Variant of the gnuplot constructor for use with do syntax. Makes sure to close the channel (and thereby) the process.

GnuplotLite.gnuplot — Method

gnuplot(persist = true, echo = false)

The main constructor for a Gnuplot instance. This starts the gnuplot process and creates a channel to which commands can be sent.

The persist argument make sure that any interactive window remains open, even if we have closed the pipe to the Gnuplot process.

The echo argument is for debugging. If enabled, all commands send to this instance are also echoed to stdout.

GnuplotLite.save — Method

 save(path::String)

Create a phony Gnuplot instance, without the underlying process attached. Instead, write messages to file. You can then later turn the file into a plot by running it with Gnuplot from the command line, or run a pipeline from Julia.

GnuplotLite.send — Method

 send(::String)

Send a literal message to a Gnuplot instance. This is supposed to be used using a pipe (|>) operator.

Returns a Msg.

GnuplotLite.send — Method

 send(::Pair{String,Matrix{T}}) where T <: Number

Send a matrix to Gnuplot, storing it in a variable. This variant assumes the matrix gets integer indices from [0..N-1], as described in the Gnuplot documentation for uniform matrices.

Returns a Msg.

GnuplotLite.send — Method

 send(::Pair{String,@NamedTuple{x::T,y::T,z::U}}) where
    {T <: AbstractVector{<:Real}, U <: AbstractMatrix{<:Real}}

Send a matrix to Gnuplot, storing it in a variable. This variant also sends axis information along with the matrix, as described in the Gnuplot documentation for nonuniform matrices.

Returns a Msg.

Example:

    x = y = -1:0.1:1
    z = x.^2 .- y'.^2
    gnuplot() do gp
        gp |> 
            send("data" => (x=collect(x), y=collect(y), z=z)) |>
            send("splot $data w l")
    end